* pokt/main:
  [RelayMiner] Supplier rate limiting (#895)
  [Tokenomics] refactor: `MintAllocationApplication` var usage to param usage  (#918)
  [Tokenomics] feat: add `mint_allocation_application` param to tokenomics module (#917)
  [Tokenomics] refactor: `MintAllocationSourceOwner` var usage to param usage (#916)
  [Tokenomics] feat: add `mint_allocation_source_owner` param to tokenomics module (#915)
  [Tokenomics] refactor: `MintAllocationSupplier` var usage to param usage (#914)
  [Tokenomics] feat: add `mint_allocation_supplier` param to tokenomics module (#913)
  [Tokenomics] refactor: `MintAllocationProposer` var usage to param usage (#912)
  [Tokenomics] feat: add `mint_allocation_proposer` param to tokenomics module (#911)
  [Tokenomics] refactor: `MintAllocationDao` var usage to param usage (#910)
  [Tokenomics] Add `mint_allocation_dao` tokenomics module param (#909)
  [quick-pr] change full-node.sh script links (#920)
  [Supplier] refactor: `NumSuppliersPerSession` var usage to param usage (#908)
  [Session Params] Add `num_suppliers_per_session` param to session module (#907)
  [Session Params] Add `MsgUpdateParam` to session module (#906)
  Updated some compiled proto files
  [TODOs] Update #2 to `TODO_BETA` to only reflect ACTUAL blockers (#900)
  [Workflow] Enforce `TODO_UPNEXT` comments to have an associated username (#896)

+++ x/proof/keeper/msg_server_update_param_test.go
@@ -16,7 +16,7 @@ import (
 )
 
 func TestMsgUpdateParam_UpdateProofRequestProbabilityOnly(t *testing.T) {
-	var expectedProofRequestProbability float32 = 0.1
+	var expectedProofRequestProbability float64 = 0.1
 
 	// Set the parameters to their default values
 	k, msgSrv, ctx := setupMsgServer(t)
